Test if cache exists

+++ travis_scripts/install_miniconda.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
-chmod +x miniconda.sh
-./miniconda.sh -b
+if test ! -e /home/travis/miniconda; then
+    wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
+    chmod +x miniconda.sh
+    ./miniconda.sh -b
+fi
